cpp 01 / ex01 finish

+++ ex0/megaphone.cpp
@@ -6,7 +6,7 @@
 /*   By: ekordi <ekordi@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 13:53:14 by ekordi            #+#    #+#             */
-/*   Updated: 2023/11/15 13:56:05 by ekordi           ###   ########.fr       */
+/*   Updated: 2024/01/16 13:29:54 by ekordi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int main(int argc, char **argv)
 		std::cout << "Invalid number of arguments" << std::endl;
 	else
 	{
-		for (size_t i = 1; i < argc; i++)
+		for (int i = 1; i < argc; i++)
 		{
 			for (size_t n = 0; argv[i][n]; n++)
 				std::cout << (char)toupper(argv[i][n]);
